+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @bakit/utils
 
+## 2.1.0
+
+### Minor Changes
+
+- @bakit/rest:
+  - Migrated interfaces to class based
+  - Simplifed REST into REST and RESTProxy
+  - Enhanced error and rate limit handling
+  - Added more events for REST and RESTProxy
+
+  @bakit/utils:
+  - Removed `createQueue` function and export `Queue` class from `p-queue`
+  - Removed `createEventBus` function.
+
 ## 2.0.0
 
 ### Patch Changes